fixes upgrades
Currently upgrades fail with:
ERROR: The running command stopped because the preference variable "ErrorActionPreference" or common parameter is set to Stop: Cannot create a file when that file already exists.

This happens because Move-Item can not overwrite directories, only files, and gives the above error when asked to do such. (Dumb huh?) So having the Move-Item ignore this error fixes the problem.